+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # calour changelog
 
+## Version 2023.4.22
+
+Bug Fixes:
+* update the Experiment.plot() function so works also with matplotlib > 3.7
+
+## Version 2022.9.8
+
+Other changes:
+* Add type hints to function return values
+* Create Experiment.info dict and populate with md5s if not provided when creating a new experiment
 
 ## Version 2022.7.1
 Incompatible changes: